% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingr.R
\docType{package}
\name{lingr}
\alias{lingr}
\alias{lingr-package}
\title{lingr: Template For Linguistics Articles.}
\description{
This package provides a basic RMarkdown template for linguisics articles
(internal use).
}
