% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lingr_article}
\alias{lingr_article}
\title{Lingr custom format.}
\usage{
lingr_article(...)
}
\arguments{
\item{...}{Arguments passed to \link[rmarkdown]{pdf_document}.}
}
\description{
A custom format for linguistics articles.
}
