% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lingr_resources}
\alias{lingr_resources}
\title{Get files from the resources directory}
\usage{
lingr_resources(path)
}
\description{
Utility function to get files from the resources directory of the package.
}
\keyword{internal}
