% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lingr_article}
\alias{lingr_article}
\title{Lingr custom format.}
\usage{
lingr_article(...)
}
\arguments{
\item{...}{Arguments passed to \link[rmarkdown]{pdf_document}.}
}
\description{
A custom format for linguistics articles. It provides the standard pandoc
YAML parameter, plus three new parameters: \code{version} for the document's
version, \code{version-date} for the date of the document's version (these parameters must be included as subkeys in \code{params}), and
\code{hide-version} which does not print the version on the document if set
to \code{true}.
}
